refactor: update imports to use isExecaErrorWithErrorCode from util module
