+++ src/util/index.ts
@@ -1,0 +1,1 @@
+export { isExecaError, isExecaErrorWithErrorCode } from "./isExecaError.js";
