+++ src/util/isExecaError.ts
